(WIP :construction:) - starting to pull out all types stuff
