--- src/common/source-files.ts
@@ -1,6 +1,0 @@
-// Dependencies:
-import * as ts from 'typescript';
-
-export interface SourceFiles {
-    [path: string]: ts.SourceFile
-};
