--- src/enums/all-enums.ts
@@ -1,6 +1,0 @@
-// Dependencies:
-import { FileEnums } from './file-enums';
-
-export interface AllEnums {
-    [path: string]: FileEnums
-}
