--- src/enums/enum-value.ts
@@ -1,5 +1,0 @@
-export class EnumValue {
-    constructor (
-        public value: any
-    ) { }
-}
