--- src/enums/file-enums.ts
@@ -1,6 +1,0 @@
-// Dependencies:
-import { EnumValue } from './enum-value';
-
-export interface FileEnums {
-    [name: string]: Array<EnumValue>
-}
