--- src/requests/imports/import.ts
@@ -1,6 +1,0 @@
-export class Import {
-    constructor (
-        public name: string,
-        public path: string
-    ) { }
-}
