--- src/requests/stack/position.ts
@@ -1,12 +1,0 @@
-// Dependencies:
-import { normalisePath } from '../normalise-path';
-
-export class Position {
-    constructor (
-      public column: number,
-      public line: number,
-      public path: string
-    ) {
-        this.path = normalisePath(this.path);
-    }
-}
