--- src/types/all-types.ts
@@ -1,6 +1,0 @@
-// Dependencies:
-import { FileTypes } from './file-types';
-
-export interface AllTypes {
-    [path: string]: FileTypes
-}
