--- src/types/file-types.ts
@@ -1,6 +1,0 @@
-// Dependencies:
-import { TypeProperty } from './type-property';
-
-export interface FileTypes {
-    [name: string]: Array<TypeProperty>
-}
