+++ test/ineeda.spec.ts
@@ -31,26 +31,6 @@ describe('ineeda:', () => {
         it('should create a mock of an interface', () => {
             let horse: IHorse = ineeda<IHorse>();
             expect(horse).to.not.be.undefined;
-        })
-
-        it('should set a property: any to an empty object', () => {
-            expect(hero.currentTool).to.deep.equal({});
-        });
-
-        it('should set a property: array to an empty array', () => {
-            expect(hero.tools).to.deep.equal([]);
-        });
-
-        it('should set a property: boolean to false', () => {
-            expect(hero.isBrave).to.equal(false);
-        });
-
-        it('should set a property: number to 0', () => {
-            expect(hero.age).to.equal(0);
-        });
-
-        it('should set a property: string to an empty string', () => {
-            expect(hero.name).to.equal('');
         });
 
         it('should set a getter for a nested complex object', () => {
@@ -70,33 +50,6 @@ describe('ineeda:', () => {
 
             expect(hero.weapon.sharpen).to.have.been.called;
         });
-
-        it('should set the lowest value for an Enum', () => {
-            expect(hero.species).to.equal(0);
-            expect(hero.weapon.type).to.equal(3);
-        });
-
-        it('should throw an error when mocking a value of an unknown type', () => {
-            expect(() => {
-                let promise = ineeda<Promise<any>>();
-            }).to.throw(`
-        Could not mock <Promise>.
-            This probably means there was no type information available for <Promise>.
-            Either add type information, or call \`ineeda<Promise>({ proxy: true });\`
-    `);
-        });
-
-        it('should throw an error when mocking a property of an unknown type', () => {
-            let hero = ineeda<Hero>();
-
-            expect(() => {
-                console.log(hero.holdOut);
-            }).to.throw(`
-        Could not mock "holdOut" on <Hero>.
-            This probably means there was no type information available for <Promise>.
-            Either add type information, or call \`ineeda<Hero>({ proxy: true });\`
-    `);
-        });
     });
 
     describe('ineeda - instanceof:', () => {
@@ -113,34 +66,6 @@ describe('ineeda:', () => {
 
         it('should be an actual instance of the class', () => {
             expect(hero).to.be.an.instanceof(Hero);
-        });
-    });
-
-    describe('ineeda - proxy', () => {
-        let hero: Hero;
-
-        beforeEach(() => {
-            let options: IneedaOptions = { proxy: true };
-            hero = ineeda<Hero>(options);
-        });
-
-        it('should return a proxied mock when mocking a value of an unknown type', () => {
-            expect(hero.holdOut).to.not.equal(undefined);
-        });
-
-        it('should return a proxied function when mocking a function of an unknown type', () => {
-            expect(() => {
-                hero.holdOut.then()
-            }).to.throw('"Hero.holdOut.then" is not implemented.');
-        });
-
-        it('should allow you to use sinon to stub a proxied function', () => {
-            sinon.stub(hero.holdOut, 'then').returns(Promise.resolve());
-
-            return hero.holdOut.then()
-            .then(() => {
-                expect(hero.holdOut.then).to.have.been.called;
-            });
         });
     });
 });
